separate config_flow and init tests
